Move Chain Db Reps to core
