+++ chain/src/main/scala/org/bitcoins/chain/api/ChainApi.scala
@@ -1,11 +1,11 @@
 package org.bitcoins.chain.api
 
-import org.bitcoins.chain.models.{
+import org.bitcoins.core.api.chain.ChainQueryApi
+import org.bitcoins.core.api.chain.db.{
   BlockHeaderDb,
   CompactFilterDb,
   CompactFilterHeaderDb
 }
-import org.bitcoins.core.api.chain.ChainQueryApi
 import org.bitcoins.core.gcs.FilterHeader
 import org.bitcoins.core.p2p.CompactFilterMessage
 import org.bitcoins.core.protocol.BlockStamp
