+++ chain/src/main/scala/org/bitcoins/chain/validation/TipValidation.scala
@@ -1,14 +1,13 @@
 package org.bitcoins.chain.validation
 
+import org.bitcoins.chain.ChainVerificationLogger
 import org.bitcoins.chain.blockchain.Blockchain
-import org.bitcoins.chain.models.{BlockHeaderDb, BlockHeaderDbHelper}
+import org.bitcoins.chain.config.ChainAppConfig
 import org.bitcoins.chain.pow.Pow
+import org.bitcoins.core.api.chain.db.{BlockHeaderDb, BlockHeaderDbHelper}
 import org.bitcoins.core.number.UInt32
 import org.bitcoins.core.protocol.blockchain.BlockHeader
 import org.bitcoins.core.util.NumberUtil
-
-import org.bitcoins.chain.config.ChainAppConfig
-import org.bitcoins.chain.ChainVerificationLogger
 
 /**
   * Responsible for checking if we can connect two
